+++ src/components/pages/Certfication/Certification.js
@@ -2,8 +2,6 @@ import React, { useState, useEffect } from "react";
 import { useHistory } from "react-router-dom";
 import Stack from '@mui/material/Stack';
 import axios from "axios";
-import ButtonUnstyled, { buttonUnstyledClasses } from '@mui/base/ButtonUnstyled';
-import { styled } from '@mui/system';
 import { Container, Typography, TableContainer, Table, TableBody, TableHead, TableRow, TableCell, TablePagination } from "@material-ui/core";
 import { makeStyles, withStyles } from "@material-ui/core/styles";
 import "../../../../src/global.css";
@@ -15,40 +13,8 @@ import PreviewIcon from '@mui/icons-material/Preview';
 import Filter from '../../Layout/FilterSearchBar/Filter';
 import { useSelector, useDispatch } from 'react-redux';
 import { updateName, updateTechStack } from '../../store/certificationFilter';
+import CustomButton from "../../common/customButton";
 
-const CustomButtonRoot = styled('button')`
-  background-color: #007fff;
-  padding: 15px 20px;
-  border-radius: 10px;
-  margin-bottom:0px;
-  color: #fff;
-  font-weight: 600;
-  font-family: Helvetica, Arial, sans-serif;
-  font-size: 14px;
-  transition: all 200ms ease;
-  cursor: pointer;
-  box-shadow: 0 4px 20px 0 rgba(61, 71, 82, 0.1), 0 0 0 0 rgba(0, 127, 255, 0);
-  border: none;
-
-  &:hover {
-    background-color: #0059b2;
-  }
-
-  &.${buttonUnstyledClasses.active} {
-    background-color: #004386;
-  }
-
-  &.${buttonUnstyledClasses.focusVisible} {
-    box-shadow: 0 4px 20px 0 rgba(61, 71, 82, 0.1), 0 0 0 5px rgba(0, 127, 255, 0.5);
-    outline: none;
-  }
-
-  &.${buttonUnstyledClasses.disabled} {
-    opacity: 0.5;
-    cursor: not-allowed;
-    box-shadow: 0 0 0 0 rgba(0, 127, 255, 0);
-  }
-`;
 const useStyles = makeStyles((theme) => ({
     root: {
         minWidth: "81vw",
@@ -95,10 +61,6 @@ const StyledTableCell = withStyles((theme) => ({
         fontSize: 14,
     },
 }))(TableCell);
-
-function CustomButton(props) {
-    return <ButtonUnstyled {...props} component={CustomButtonRoot} />;
-}
 
 export default function Certification(props) {
     const [page, setPage] = useState(0);
