+++ src/components/pages/Poc/poc.js
@@ -2,8 +2,7 @@ import React, { useState, useEffect } from 'react';
 import { useHistory } from 'react-router-dom';
 import Stack from '@mui/material/Stack';
 import axios from 'axios';
-import ButtonUnstyled, { buttonUnstyledClasses } from '@mui/base/ButtonUnstyled';
-import { styled } from '@mui/system';
+import CustomButton from '../../common/customButton';
 import {
 	Container,
 	Typography,
@@ -28,39 +27,6 @@ import { updateName, updateTechStack } from '../../store/pocFilter';
 import { getApiUrl } from '../../common/helper';
 import FormDialog from '../../common/formDialog';
 
-const CustomButtonRoot = styled('button')`
-	background-color: #007fff;
-	padding: 15px 20px;
-	border-radius: 10px;
-	margin-bottom: 0px;
-	color: #fff;
-	font-weight: 600;
-	font-family: Helvetica, Arial, sans-serif;
-	font-size: 14px;
-	transition: all 200ms ease;
-	cursor: pointer;
-	box-shadow: 0 4px 20px 0 rgba(61, 71, 82, 0.1), 0 0 0 0 rgba(0, 127, 255, 0);
-	border: none;
-
-	&:hover {
-		background-color: #0059b2;
-	}
-
-	&.${buttonUnstyledClasses.active} {
-		background-color: #004386;
-	}
-
-	&.${buttonUnstyledClasses.focusVisible} {
-		box-shadow: 0 4px 20px 0 rgba(61, 71, 82, 0.1), 0 0 0 5px rgba(0, 127, 255, 0.5);
-		outline: none;
-	}
-
-	&.${buttonUnstyledClasses.disabled} {
-		opacity: 0.5;
-		cursor: not-allowed;
-		box-shadow: 0 0 0 0 rgba(0, 127, 255, 0);
-	}
-`;
 const useStyles = makeStyles((theme) => ({
 	root: {
 		minWidth: '81vw',
@@ -107,10 +73,6 @@ const StyledTableCell = withStyles((theme) => ({
 		fontSize: 14,
 	},
 }))(TableCell);
-
-function CustomButton(props) {
-	return <ButtonUnstyled {...props} component={CustomButtonRoot} />;
-}
 
 export default function POC(props) {
 	const [page, setPage] = useState(0);
